deep-research/Manual Maestro Inyección Dependencias FastAPI/fastapi-service-master/pyproject.toml
[project]
name = "fastapi-service-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
]
